fix: 시간표 구조 컴포넌트화 진행
