+++ src/utils/zustand/myLecture.ts
@@ -1,40 +1,40 @@
+/* eslint-disable @typescript-eslint/no-unused-vars */
 /* eslint-disable no-useless-return */
+import { LectureInfo } from 'interfaces/Lecture';
+import createSelectors from 'utils/zustand/createSelector';
+import { create } from 'zustand';
 
-// import { LectureInfo } from 'interfaces/Lecture';
-// import createSelectors from 'utils/zustand/createSelector';
-// import { create } from 'zustand';
+type SelectedSemester = string;
 
-// type SelectedSemester = string;
+const temptokenState = '';
 
-// const temptokenState = '';
+type State = {
+  myLectures: LectureInfo[];
+};
 
-// type State = {
-//   myLectures: LectureInfo[];
-// };
+type Actions = {
+  setMyLecture: (newValue: LectureInfo) => void;
+};
 
-// type Actions = {
-//   setMyLecture: (newValue: LectureInfo) => void;
-// };
+const initialState: State = {
+  myLectures: [],
+};
 
-// const initialState: State = {
-//   myLectures: null,
-// };
-
-// function waitForTruthyValue<T>(getValue: () => T, timeout = 1000): Promise<T> {
-//   return new Promise((resolve) => {
-//     function endSetTimeoutWhenValueNonNullable() {
-//       setTimeout(() => {
-//         const value = getValue();
-//         if (value) {
-//           resolve(value);
-//         } else {
-//           endSetTimeoutWhenValueNonNullable();
-//         }
-//       }, timeout);
-//     }
-//     endSetTimeoutWhenValueNonNullable();
-//   });
-// }
+function waitForTruthyValue<T>(getValue: () => T, timeout = 1000): Promise<T> {
+  return new Promise((resolve) => {
+    function endSetTimeoutWhenValueNonNullable() {
+      setTimeout(() => {
+        const value = getValue();
+        if (value) {
+          resolve(value);
+        } else {
+          endSetTimeoutWhenValueNonNullable();
+        }
+      }, timeout);
+    }
+    endSetTimeoutWhenValueNonNullable();
+  });
+}
 
 // const myLectureAddLecture = create<State & Actions>()((set, get) => ({
 //   ...initialState,
